src/main/webapp/app/entities/galeria-imagen/delete/galeria-imagen-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { GaleriaImagenService } from '../service/galeria-imagen.service';

import { GaleriaImagenDeleteDialogComponent } from './galeria-imagen-delete-dialog.component';

describe('GaleriaImagen Management Delete Component', () => {
  let comp: GaleriaImagenDeleteDialogComponent;
  let fixture: ComponentFixture<GaleriaImagenDeleteDialogComponent>;
  let service: GaleriaImagenService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [GaleriaImagenDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(GaleriaImagenDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(GaleriaImagenDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(GaleriaImagenService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
